Assignment 1: Introduction to data structures labs
This one is as simple as it gets. It was mostly for us to get used to the zybooks website where we worked on these assignments.
We were assigned to implement the CallFunctionNamed() to handle the 2 possible cases provided in the LabPrinter.h file, along with an "Unkown function" exception if the other 2 if checks never returned out of the function.