Assignment 2: Binary search template function
This lesson was all about Searching and Algorithm Analysis. In this assignment we first needed to implement the BinarySearch() function in the Searcher.h file. BinarySearch() performs a binary search on the sorted array (first parameter) for the key (third parameter). BinarySearch() returns the key's index if found, -1 if not found.
Very simple binary search assignment.