Assignment 3: Natural merge sort
This lesson was all about Sorting Algorithms. Here we are given a merge function that is a merge sort algorithm. The merge sort algorithm recursively divides the array in half until an array with one element is reached. We have to implement a natural merge sort, a variant of merge sort. Natural merge sorts find already-sorted runs of elements and merge the runs together.
First we need to implement GetSortedRunLength(). It returns the number of array elements sorted in ascending order, starting at startIndex and ending either at the end of the sorted run, or the end of the array, whichever comes first. The function returns 0 if startIndex is out of bounds.
Second we implement our natural merge sort.
As implemented, NaturalMergeSort() takes a pointer to an array of integers as well as the array's length.
First: The index i starts at 0
Second: Gets the length of the first sorted run, starting at i
Returns if the first run's length equals the array length
If the first run ends at the array's end, reassign i=0 and repeat second step
Third: Get the length of the second sorted run, starting immediately after the first
Fourth: Merge the two runs with the provided Merge() function
Fifth: Reassign i with the first index after the second run, or 0 if the second run ends at the array's end
Sixth: Go to second step